dodavanje node-cron, brisanje postova dnevno

+++ src/graphql/resolvers/post.js
@@ -19,14 +19,14 @@ const resolvers = {
       }));
       return postsNew;
     },
-    getClientPosts: async (_, { email }) => {
-      console.log("getClientPosts(" + email + ")");
+    getClientPosts: async (_, { email }, { client }) => {
+      if (!client) {
+        return null;
+      }
       const posts = await Post.find()
         .sort({ createdAt: -1 })
         .populate("author");
-      //   console.log(posts);
-      //   console.log(posts?.[0]?.id);
-      //   console.log(posts[0].author._id.toString());
+
       return posts
         .filter((post) => post?.author?.email === email)
         .map((post) => ({
@@ -37,8 +37,10 @@ const resolvers = {
         }));
     },
 
-    getPostById: async (_, { id }) => {
-      console.log("get post by id: " + id);
+    getPostById: async (_, { id }, { client }) => {
+      if (!client) {
+        return null;
+      }
       const post = await Post.findById({
         _id: new mongoose.Types.ObjectId(id),
       }).populate("author");
@@ -50,11 +52,15 @@ const resolvers = {
     },
   },
   Mutation: {
-    createNewPost: async (_, { newPostInput: { content, author: email } }) => {
-      let postAuthor = await Client.find();
-      //   const postAuthor2 = await Client.find({ email });
-
-      postAuthor = postAuthor.find((auhtor) => auhtor.email === email);
+    createNewPost: async (
+      _,
+      { newPostInput: { content, author: email } },
+      { client }
+    ) => {
+      if (!client) {
+        return null;
+      }
+      let postAuthor = await Client.findOne({ email });
 
       if (!content.length) {
         throw new ApolloError(`Post field can't be empty!`, "EMPTY_POST_FIELD");
@@ -64,17 +70,17 @@ const resolvers = {
         content,
         author: postAuthor,
       });
-      console.log("napravlje");
       const res = await post.save();
 
-      console.log("dodana u bazu");
       return {
         id: res.id,
         ...res._doc,
       };
     },
-    deletePost: async (_, { deletePost: { id } }) => {
-      console.log("deleting");
+    deletePost: async (_, { deletePost: { id } }, { client }) => {
+      if (!client) {
+        return null;
+      }
       //finding by id and deleting it
       await Post.deleteOne({
         _id: new mongoose.Types.ObjectId(id),
@@ -82,8 +88,10 @@ const resolvers = {
 
       return "Successfully deleted a post!";
     },
-    editPost: async (_, { editPostInput: { id, content } }) => {
-      console.log("editing post");
+    editPost: async (_, { editPostInput: { id, content } }, client) => {
+      if (!client) {
+        return null;
+      }
       if (!content.length) {
         throw new ApolloError(`Post field can't be empty!`, "EMPTY_POST_FIELD");
       }
@@ -94,8 +102,6 @@ const resolvers = {
 
       const res = await post.save();
 
-      console.log("ID : " + id);
-      console.log("content: ", content);
       return {
         id: res.id,
         ...res._doc,
